Info.lua

+++ XXTExplorer/Detail/Terminal/XXTELuaVModel.m
@@ -13,31 +13,16 @@
 
 #import "XXTELuaVModel.h"
 #import "XXTEAppDelegate.h"
+#import "LuaNSValue.h"
 
-static NSString * const kXXTerminalFakeHandlerStandardOutput = @"kXXTerminalFakeHandlerStandardOutput-%@.pipe";
-static NSString * const kXXTerminalFakeHandlerStandardError = @"kXXTerminalFakeHandlerStandardError-%@.pipe";
-static NSString * const kXXTerminalFakeHandlerStandardInput = @"kXXTerminalFakeHandlerStandardInput-%@.pipe";
+static NSString * const XXTETerminalHandlerOutput = @"TerminalOutput-%@.pipe";
+static NSString * const XXTETerminalHandlerError = @"TerminalError-%@.pipe";
+static NSString * const XXTETerminalHandlerInput = @"TerminalInput-%@.pipe";
 
 static jmp_buf buf;
 static BOOL _running = NO;
 
-void luaL_setPath(lua_State* L, const char *key, const char *path)
-{
-    lua_getglobal(L, "package");
-    lua_getfield(L, -1, key); // get field "path" from table at top of stack (-1)
-    const char *origPath = lua_tostring(L, -1); // grab path string from top of stack
-    NSString *strPath = [[NSString alloc] initWithUTF8String:path];
-    NSString *strOrigPath = [[NSString alloc] initWithUTF8String:origPath];
-    strOrigPath = [strOrigPath stringByAppendingString:@";"];
-    strOrigPath = [strOrigPath stringByAppendingString:strPath];
-    strOrigPath = [strOrigPath stringByAppendingString:@";"];
-    lua_pop(L, 1); // get rid of the string on the stack we just pushed on line 5
-    lua_pushstring(L, [strOrigPath UTF8String]); // push the new one
-    lua_setfield(L, -2, key); // set the field "path" in table at -2 with value at top of stack
-    lua_pop(L, 1); // get rid of package table from top of stack
-}
-
-void luaL_terminate(lua_State *L, lua_Debug *ar)
+void lua_terminate(lua_State *L, lua_Debug *ar)
 {
     if (!_running) {
         longjmp(buf, 1);
@@ -49,7 +34,7 @@ void luaL_terminate(lua_State *L, lua_Debug *ar)
 @end
 
 @implementation XXTELuaVModel {
-    lua_State *luaState;
+    lua_State *L;
 }
 
 - (instancetype)init {
@@ -61,19 +46,30 @@ void luaL_terminate(lua_State *L, lua_Debug *ar)
 
 - (void)setup {
     @synchronized (self) {
-        NSString *stdoutHandlerPath = [NSTemporaryDirectory() stringByAppendingPathComponent:[NSString stringWithFormat:kXXTerminalFakeHandlerStandardOutput, [NSUUID UUID]]];
+        if (!L) {
+            L = luaL_newstate();
+            NSAssert(L, @"LuaVM: not enough memory.");
+            lua_sethook(L, &lua_terminate, LUA_MASKLINE, 1);
+            luaL_openlibs(L);
+        }
+    }
+}
+
+- (void)setFakeIOEnabled:(BOOL)enabled {
+    if (enabled) {
+        NSString *stdoutHandlerPath = [NSTemporaryDirectory() stringByAppendingPathComponent:[NSString stringWithFormat:XXTETerminalHandlerOutput, [NSUUID UUID]]];
         unlink(stdoutHandlerPath.UTF8String);
         FILE *stdoutHandler = fopen(stdoutHandlerPath.UTF8String, "wb+");
         NSAssert(stdoutHandler, @"Cannot create stdout handler");
         self.stdoutHandler = stdoutHandler;
         
-        NSString *stderrHandlerPath = [NSTemporaryDirectory() stringByAppendingPathComponent:[NSString stringWithFormat:kXXTerminalFakeHandlerStandardError, [NSUUID UUID]]];
+        NSString *stderrHandlerPath = [NSTemporaryDirectory() stringByAppendingPathComponent:[NSString stringWithFormat:XXTETerminalHandlerError, [NSUUID UUID]]];
         unlink(stderrHandlerPath.UTF8String);
         FILE *stderrHandler = fopen(stderrHandlerPath.UTF8String, "wb+");
         NSAssert(stderrHandler, @"Cannot create stderr handler");
         self.stderrHandler = stderrHandler;
         
-        NSString *stdinHandlerPath = [NSTemporaryDirectory() stringByAppendingPathComponent:[NSString stringWithFormat:kXXTerminalFakeHandlerStandardInput, [NSUUID UUID]]];
+        NSString *stdinHandlerPath = [NSTemporaryDirectory() stringByAppendingPathComponent:[NSString stringWithFormat:XXTETerminalHandlerInput, [NSUUID UUID]]];
         unlink(stdinHandlerPath.UTF8String);
         if (mkfifo(stdinHandlerPath.UTF8String, S_IRWXU) >= 0)
         {
@@ -91,12 +87,25 @@ void luaL_terminate(lua_State *L, lua_Debug *ar)
         }
         
         fakeio(self.stdinReadHandler, self.stdoutHandler, self.stderrHandler);
+    } else {
+        fakeio(stdin, stdout, stderr);
+        chdir("/");
         
-        if (!luaState) {
-            luaState = luaL_newstate();
-            NSAssert(luaState, @"not enough memory");
-            lua_sethook(luaState, &luaL_terminate, LUA_MASKLINE, 1);
-            luaL_openlibs(luaState);
+        if (self.stdoutHandler) {
+            fclose(self.stdoutHandler);
+            self.stdoutHandler = nil;
+        }
+        if (self.stderrHandler) {
+            fclose(self.stderrHandler);
+            self.stderrHandler = nil;
+        }
+        if (self.stdinReadHandler) {
+            fclose(self.stdinReadHandler);
+            self.stdinReadHandler = nil;
+        }
+        if (self.stdinWriteHandler) {
+            fclose(self.stdinWriteHandler);
+            self.stdinWriteHandler = nil;
         }
     }
 }
@@ -111,10 +120,10 @@ void luaL_terminate(lua_State *L, lua_Debug *ar)
     _running = running;
     if (!running)
     {
-        char *emptyBuf = malloc(8192 * sizeof(char));
+        char *emptyBuf = malloc(8192 * sizeof(char)); // malloc
         memset(emptyBuf, 0x0a, 8192);
         write(fileno(self.stdinWriteHandler), emptyBuf, 8192);
-        free(emptyBuf);
+        free(emptyBuf); // free
     }
     if (_delegate && [_delegate respondsToSelector:@selector(virtualMachineDidChangedState:)])
     {
@@ -129,8 +138,8 @@ void luaL_terminate(lua_State *L, lua_Debug *ar)
     NSString *sPath = [NSString stringWithFormat:@"%@", [dirPath stringByAppendingPathComponent:@"?.lua"]];
     NSString *cPath = [NSString stringWithFormat:@"%@", [dirPath stringByAppendingPathComponent:@"?.so"]];
     @synchronized (self) {
-        luaL_setPath(luaState, "path", sPath.UTF8String);
-        luaL_setPath(luaState, "cpath", cPath.UTF8String);
+        lua_setPath(L, "path", sPath.UTF8String);
+        lua_setPath(L, "cpath", cPath.UTF8String);
     }
 }
 
@@ -140,16 +149,22 @@ void luaL_terminate(lua_State *L, lua_Debug *ar)
 {
     [self setCurrentPath:[path stringByDeletingLastPathComponent]];
     const char *cString = [path UTF8String];
-    int load_stat = luaL_loadfile(luaState, cString);
-    return checkCode(luaState, load_stat, error);
+    int load_stat = luaL_loadfile(L, cString);
+    if (!checkCode(L, load_stat, error)) {
+        return NO;
+    }
+    return YES;
 }
 
 - (BOOL)loadBufferFromString:(NSString *)string
                        error:(NSError **)error
 {
     const char *cString = [string UTF8String];
-    int load_stat = luaL_loadbufferx(luaState, cString, strlen(cString), "", 0);
-    return checkCode(luaState, load_stat, error);
+    int load_stat = luaL_loadbufferx(L, cString, strlen(cString), "", 0);
+    if (!checkCode(L, load_stat, error)) {
+        return NO;
+    }
+    return YES;
 }
 
 #pragma mark - pcall
@@ -158,13 +173,12 @@ void luaL_terminate(lua_State *L, lua_Debug *ar)
     self.running = YES;
     int load_stat = 0;
     if (!setjmp(buf)) {
-        load_stat = lua_pcall(luaState, 0, 0, 0);
+        load_stat = lua_pcall(L, 0, 0, 0);
         self.running = NO;
-        BOOL loadResult = checkCode(luaState, load_stat, error);
-        if (loadResult) {
-            
+        if (!checkCode(L, load_stat, error)) {
+            return NO;
         }
-        return loadResult;
+        return YES;
     } else {
         if (error != nil) {
             *error = [NSError errorWithDomain:kXXTELuaVModelErrorDomain code:-1 userInfo:@{ NSLocalizedFailureReasonErrorKey: NSLocalizedString(@"Thread terminated.", nil) }];
@@ -176,31 +190,10 @@ void luaL_terminate(lua_State *L, lua_Debug *ar)
 #pragma mark - memory
 
 - (void)dealloc {
-    if (luaState) {
-        lua_close(luaState);
-        luaState = NULL;
+    if (L) {
+        lua_close(L);
+        L = NULL;
     }
-    
-    fakeio(stdin, stdout, stderr);
-    chdir("/");
-    
-    if (self.stdoutHandler) {
-        fclose(self.stdoutHandler);
-        self.stdoutHandler = nil;
-    }
-    if (self.stderrHandler) {
-        fclose(self.stderrHandler);
-        self.stderrHandler = nil;
-    }
-    if (self.stdinReadHandler) {
-        fclose(self.stdinReadHandler);
-        self.stdinReadHandler = nil;
-    }
-    if (self.stdinWriteHandler) {
-        fclose(self.stdinWriteHandler);
-        self.stdinWriteHandler = nil;
-    }
-    
 #ifdef DEBUG
     NSLog(@"- [XXTELuaVModel dealloc]");
 #endif
